+++ src/Phan/Language/UnionType.php
@@ -22,6 +22,7 @@ use Phan\Language\Type\MixedType;
 use Phan\Language\Type\MultiType;
 use Phan\Language\Type\NullType;
 use Phan\Language\Type\StaticType;
+use Phan\Language\Type\StringType;
 use Phan\Language\Type\TemplateType;
 use Phan\Language\Type\TrueType;
 use ast\Node;
@@ -1529,6 +1530,7 @@ class UnionType implements \Serializable
      * type.
      *
      * @return \Generator
+     * @phan-return \Generator<FullyQualifiedClassName>
      *
      * A list of class FQSENs representing the non-native types
      * associated with this UnionType
@@ -1554,6 +1556,10 @@ class UnionType implements \Serializable
             }
             // Get the class FQSEN
             $class_fqsen = $class_type->asFQSEN();
+            if (!($class_fqsen instanceof FullyQualifiedClassName)) {
+                // Should be impossible, but skip to satisfy the type checker
+                continue;
+            }
 
             if ($class_type->isStaticType()) {
                 if (!$context->isInClassScope()) {
@@ -1892,11 +1898,84 @@ class UnionType implements \Serializable
     }
 
     /**
+     * Takes "a|b[]|c|d[]|e|Traversable<f,g>" and returns "int|string|f"
+     *
+     * Takes "array{field:int,other:stdClass}" and returns "string"
+     *
+     * @param CodeBase $code_base (for detecting the iterable value types of `class MyIterator extends Iterator`)
+     *
+     * @return UnionType
+     */
+    public function iterableKeyUnionType(CodeBase $code_base) : UnionType
+    {
+        // This is frequently called, and has been optimized
+        $builder = new UnionTypeBuilder();
+        $type_set = $this->type_set;
+        foreach ($type_set as $type) {
+            $element_type = $type->iterableKeyUnionType($code_base);
+            if ($element_type === null) {
+                // Does not have iterable values
+                continue;
+            }
+            $builder->addUnionType($element_type);
+        }
+
+        return $builder->getUnionType();
+    }
+
+    /**
+     * Takes "a|b[]|c|d[]|e|Traversable<f,g>" and returns "b|d|g"
+     *
+     * Takes "array{field:int,other:string}" and returns "int|string"
+     *
+     * @param CodeBase $code_base (for detecting the iterable value types of `class MyIterator extends Iterator`)
+     *
+     * @return UnionType
+     */
+    public function iterableValueUnionType(CodeBase $code_base) : UnionType
+    {
+        // This is frequently called, and has been optimized
+        $builder = new UnionTypeBuilder();
+        $type_set = $this->type_set;
+        foreach ($type_set as $type) {
+            $element_type = $type->iterableValueUnionType($code_base);
+            if ($element_type === null) {
+                // Does not have iterable values
+                continue;
+            }
+            $builder->addUnionType($element_type);
+        }
+
+        static $array_type_nonnull = null;
+        static $array_type_nullable = null;
+        static $mixed_type = null;
+        static $null_type = null;
+        if ($array_type_nonnull === null) {
+            $array_type_nonnull = ArrayType::instance(false);
+            $array_type_nullable = ArrayType::instance(true);
+            $mixed_type = MixedType::instance(false);
+            $null_type = NullType::instance(false);
+        }
+
+        // If array is in there, then it can be any type
+        if (\in_array($array_type_nonnull, $type_set, true)) {
+            $builder->addType($mixed_type);
+            $builder->addType($null_type);
+        } elseif (\in_array($mixed_type, $type_set, true)
+            || \in_array($array_type_nullable, $type_set, true)
+        ) {
+            // Same for mixed
+            $builder->addType($mixed_type);
+        }
+
+        return $builder->getUnionType();
+    }
+
+    /**
      * Takes "a|b[]|c|d[]|e" and returns "b|d"
      * Takes "array{field:int,other:string}" and returns "int|string"
      *
      * @return UnionType
-     * The subset of types in this
      */
     public function genericArrayElementTypes() : UnionType
     {
@@ -1946,6 +2025,8 @@ class UnionType implements \Serializable
      *
      * @return UnionType
      * The subset of types in this
+     *
+     * TODO: Add a variant that will convert mixed to array<int,mixed> instead of array?
      */
     public function elementTypesToGenericArray(int $key_type) : UnionType
     {
@@ -2254,6 +2335,7 @@ class UnionType implements \Serializable
                 self::convertToTypeSetWithNormalizedNonNullableBools($builder);
             }
         }
+        // TODO: Convert array|array{} to array?
         return $builder->getUnionType();
     }
 
@@ -2475,6 +2557,40 @@ class UnionType implements \Serializable
             // Swallow "Cannot find class", go on to emit issue
         }
         return false;
+    }
+
+    public function asGeneratorTemplateType() : Type
+    {
+        $fallback_values = UnionType::empty();
+        $fallback_keys = UnionType::empty();
+
+        foreach ($this->getTypeSet() as $type) {
+            if ($type->isGenerator()) {
+                if ($type->hasTemplateParameterTypes()) {
+                    return $type;
+                }
+            }
+            // TODO: support Iterator<T> or Traversable<T> or iterable<T>
+            if ($type instanceof GenericArrayType) {
+                $fallback_values = $fallback_values->withType($type->genericArrayElementType());
+                $key_type = $type->getKeyType();
+                if ($key_type === GenericArrayType::KEY_INT) {
+                    $fallback_keys = $fallback_keys->withType(IntType::instance(false));
+                } elseif ($key_type === GenericArrayType::KEY_STRING) {
+                    $fallback_keys = $fallback_keys->withType(StringType::instance(false));
+                }
+            } elseif ($type instanceof ArrayShapeType && $type->isNotEmptyArrayShape()) {
+                $fallback_values = $fallback_values->withUnionType($type->genericArrayElementUnionType());
+                $fallback_keys = $fallback_keys->withUnionType(GenericArrayType::unionTypeForKeyType($type->getKeyType()));
+            }
+        }
+
+        $result = Type::fromFullyQualifiedString('\Generator');
+        if ($fallback_keys->typeCount() > 0 || $fallback_values->typeCount() > 0) {
+            $template_types = $fallback_keys->typeCount() > 0 ? [$fallback_keys, $fallback_values] : [$fallback_values];
+            $result = $result->fromType($result, $template_types);
+        }
+        return $result;
     }
 }